add a more flexible and minimal version of the icinga module
this flexibility is payed with redundant definitions in your
i3status.conf
